Update Moved all Model related classes into their own subproject for better organisation.
Added multiple new Classes to Model

Created two usercontrols to be used as templates for the data classes Behaviour and Target.
They were only created and are not useable yet.

Added data bindings for multiple gui elements.

Added a ValueConverter to check if an object is null.
The ValueConverter gets used to set the enabled status of the Details tab representing the currently selected SortingUnit in a ListBox.